% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_utf8.R
\name{chk_utf8}
\alias{chk_utf8}
\title{Check UTF-8 Symbol Width in IDE}
\usage{
chk_utf8(x, times, total)
}
\arguments{
\item{x}{\link[base]{character} scalar}

\item{times, total}{\link[base]{integer} scalars}
}
\description{
Check UTF-8 Symbol Width in IDE
}
\examples{
# RStudio
chk_utf8('\u8336', times = 60L, total = 100L) # Chinese
chk_utf8('\u304a', times = 60L, total = 100L) # hiragana
chk_utf8('\u30aa', times = 60L, total = 100L) # katakana
chk_utf8('\ub3c4', times = 70L, total = 101L) # hanga
chk_utf8('\U1f375', times = 60L, total = 100L) # emoji
# Positron, mac default
chk_utf8('\u8336', times = 85L, total = 141L) # Chinese
chk_utf8('\u304a', times = 85L, total = 141L) # hiragana
chk_utf8('\u30aa', times = 85L, total = 141) # katakana
chk_utf8('\U1f375', times = 85L, total = 141) # emoji
chk_utf8('\ub3c4', times = 98L, total = 141) # hanga
# RGui, mac default
chk_utf8('\u8336', times = 84L, total = 140) # Chinese
chk_utf8('\u304a', times = 84L, total = 140) # hiragana
chk_utf8('\u30aa', times = 84L, total = 140) # katakana
chk_utf8('\U1f375', times = 60L, total = 136) # emoji
chk_utf8('\ub3c4', times = 95L, total = 137) # hanga

# not recognized by any rendering
chk_utf8('\u2009', times = 100L, total = 100L) # thin space
chk_utf8('\u202f', times = 100L, total = 100L) # narrow no-break space

}
\keyword{character}
\keyword{rendered}
\keyword{unicode}
\keyword{utf8}
\keyword{visualize}
\keyword{width}
