% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ANSI_matrix.R
\name{print_ANSI_matrix}
\alias{print_ANSI_matrix}
\title{Print \link[base]{matrix} with ANSI Escape Sequences}
\usage{
print_ANSI_matrix(x)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{matrix}}
}
\value{
The function \code{\link[=print_ANSI_matrix]{print_ANSI_matrix()}} does not have a returned value.
}
\description{
Print \link[base]{matrix} with ANSI escape sequences in the body and/or \link[base]{dimnames}.
}
\seealso{
\code{colorDF::colorDF}
}
