% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8_width.R
\name{utf8_width}
\alias{utf8_width}
\title{UTF-8 Symbol Width in IDE}
\usage{
utf8_width(x, times, total)
}
\arguments{
\item{x}{\link[base]{character} scalar}

\item{times, total}{\link[base]{integer} scalars}
}
\description{
UTF-8 Symbol Width in IDE
}
\examples{
# RStudio
utf8_width('\u8336', times = 60L, total = 100L) # Chinese
utf8_width('\u304a', times = 60L, total = 100L) # hiragana
utf8_width('\u30aa', times = 60L, total = 100L) # katakana
utf8_width('\ub3c4', times = 70L, total = 101L) # hanga
utf8_width('\U1f375', times = 60L, total = 100L) # emoji
# Positron, mac default
utf8_width('\u8336', times = 85L, total = 141L) # Chinese
utf8_width('\u304a', times = 85L, total = 141L) # hiragana
utf8_width('\u30aa', times = 85L, total = 141) # katakana
utf8_width('\U1f375', times = 85L, total = 141) # emoji
utf8_width('\ub3c4', times = 98L, total = 141) # hanga
# RGui, mac default
utf8_width('\u8336', times = 84L, total = 140) # Chinese
utf8_width('\u304a', times = 84L, total = 140) # hiragana
utf8_width('\u30aa', times = 84L, total = 140) # katakana
utf8_width('\U1f375', times = 60L, total = 136) # emoji
utf8_width('\ub3c4', times = 95L, total = 137) # hanga

}
