% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_justify.R
\name{ws_justify}
\alias{ws_justify}
\title{Padded Whitespace for Left/Right Justification}
\usage{
ws_justify(x)
}
\arguments{
\item{x}{an R object of \link[base]{mode} \link[base]{character}}
}
\value{
The function \code{\link[=ws_justify]{ws_justify()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
Padded whitespace for left/right justification.
}
\examples{
ws_justify(c('a', 'abc', 'ab'))
ws_justify(matrix(c('a', 'abc', 'ab', 'abcd'), nrow = 2L))

(x = c('apple', 'tea\U1f375', '\U1f1fa\U1f1f8 and \U1f1e8\U1f1e6'))
paste0(ws_justify(x), x) |> cat(sep = '\n')
(x1 = stringi::stri_dup(x, times = 10L))
paste0(ws_justify(x1), x1) |> cat(sep = '\n')

}
\seealso{
\link[base]{format.default}
}
