% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_matrix.R
\name{cat_matrix}
\alias{cat_matrix}
\title{An Alternative Print of \link[base]{matrix} by \link[base]{cat}}
\usage{
cat_matrix(x)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{matrix}}
}
\value{
The function \code{\link[=cat_matrix]{cat_matrix()}} does not have a returned value.
}
\description{
An alternative print of \link[base]{matrix} by \link[base]{cat}.
}
