% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charwidth.R
\name{charwidth}
\alias{charwidth}
\title{An Alternative \link[base]{nchar}}
\usage{
charwidth(x)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector}}
}
\description{
An Alternative \link[base]{nchar}
}
\examples{
(x = c('tea\u8336\ub2e4\ub3c4\U1f375\u207a'))
charwidth(x)
# RGui: 10.81754 * 11 # approx 119
# RStudio: 10.21905 * 9 # approx 92L
# Positron: 10.1952 * 15 # approx 153
\dontrun{
library(stringi)
cat(stri_dup(x, 11L), stri_dup('a', 119L), sep = '\n') # RGui
cat(stri_dup(x, 9L), stri_dup('a', 92L), sep = '\n') # RStudio
cat(stri_dup(x, 15L), stri_dup('a', 153L), sep = '\n') # Positron
}

x |>
 cli::col_red() |>
 charwidth()
}
