% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charwidth.R
\name{charwidth}
\alias{charwidth}
\title{An Alternative \link[base]{nchar}}
\usage{
charwidth(x)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector}}
}
\description{
An Alternative \link[base]{nchar}
}
\examples{
(x = c('tea\u8336\ub2e4\ub3c4\U1f375\u207a'))
charwidth(x) # 10.21905
\dontrun{
10.21905 * 9 # approx 92L
library(stringi)
cat(stri_dup(x, 9L), stri_dup('a', 92L), sep = '\n')
}

}
