% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charwidth.R
\name{charwidth}
\alias{charwidth}
\title{An Alternative \link[base]{nchar}}
\usage{
charwidth(x)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector}}
}
\description{
An Alternative \link[base]{nchar}
}
\examples{
c('tea\U1f375', '\U1f1fa\U1f1f8 and \U1f1e8\U1f1e6') |>
  charwidth()

}
