% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charwidth.R
\name{charwidth}
\alias{charwidth}
\title{An Alternative \link[base]{nchar}}
\usage{
charwidth(x)
}
\arguments{
\item{x}{\link[base]{character} \link[base]{vector}}
}
\description{
An Alternative \link[base]{nchar}
}
\examples{
x10 = '\u4e00\u4e8c\u4e09\u56db\u4e94\u516d\u4e03\u516b\u4e5d\u96f6'
x20 = '1234567890'
cat(x10, '\n', x20, sep = '')
x1 = x10 |> substr(start = 1L, stop = 6L) |> stringi::stri_dup(times = 20L)
x2 = x20 |> stringi::stri_dup(times = 20L)
cat(x1, '\n', x2, sep = '')

x = c('tea\U1f375', '\U1f1fa\U1f1f8 and \U1f1e8\U1f1e6')
charwidth(x)

}
