% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_fmt_matrix.R
\name{row_fmt_matrix}
\alias{row_fmt_matrix}
\title{Format \link[base]{matrix} by Rows}
\usage{
row_fmt_matrix(x, rownm_justify = "left", justify = "right")
}
\arguments{
\item{x}{\link[base]{character} \link[base]{matrix}}

\item{rownm_justify}{\link[base]{character} scalar, default value is \code{'left'}, see the function \code{\link[=ws_fmt]{ws_fmt()}}}

\item{justify}{\link[base]{character} scalar, default value is \code{'right'}, see the function \code{\link[=ws_fmt]{ws_fmt()}}}
}
\value{
The function \code{\link[=row_fmt_matrix]{row_fmt_matrix()}} returns a \link[base]{list} of \link[base]{character} \link[base]{vector}s of the formatted \strong{rows}.
}
\description{
Format \link[base]{matrix} by rows.
}
\examples{
VADeaths |> row_fmt_matrix()

}
