% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ws_fmt.R
\name{ws_fmt}
\alias{ws_fmt}
\title{Whitespace-Padded for Left/Right Justification}
\usage{
ws_fmt(x, justify = c("left", "right"))
}
\arguments{
\item{x}{an R object of \link[base]{mode} \link[base]{character}}

\item{justify}{\link[base]{character} scalar, only options are \code{'left'} and \code{'right'}, see the function \link[base]{format.default}}
}
\value{
The function \code{\link[=ws_fmt]{ws_fmt()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
Whitespace-padded for left/right justification.
}
\examples{
ws_fmt(c('a', 'abc', 'ab'))
ws_fmt(c('a', 'abc', 'ab'), justify = 'right')

}
\seealso{
\link[base]{format.default}
}
